64bit fixes
Boots but when it plays a demo or your start a new game, you get the dreaded "Error: Z_Free: freed a pointer without ZONEID".  This will take some time.  32bit is fine as usual.

+++ r_draw.c
@@ -27,7 +27,7 @@
 static const char
 rcsid[] = "$Id: r_draw.c,v 1.4 1997/02/03 16:47:55 b1 Exp $";
 
-
+#include <stdint.h>
 #include "doomdef.h"
 
 #include "i_system.h"
@@ -461,29 +461,29 @@ void R_DrawTranslatedColumn (void)
 // Assumes a given structure of the PLAYPAL.
 // Could be read from a lump instead.
 //
-void R_InitTranslationTables (void)
+void R_InitTranslationTables(void)
 {
     int		i;
-	
-    translationtables = Z_Malloc (256*3+255, PU_STATIC, 0);
-    translationtables = (byte *)(( (int)translationtables + 255 )& ~255);
-    
+
+    translationtables = Z_Malloc(256 * 3 + 255, PU_STATIC, 0);
+    translationtables = (byte*)(((intptr_t)translationtables + 255) & ~255);
+
     // translate just the 16 green colors
-    for (i=0 ; i<256 ; i++)
+    for (i = 0; i < 256; i++)
     {
-	if (i >= 0x70 && i<= 0x7f)
-	{
-	    // map green ramp to gray, brown, red
-	    translationtables[i] = 0x60 + (i&0xf);
-	    translationtables [i+256] = 0x40 + (i&0xf);
-	    translationtables [i+512] = 0x20 + (i&0xf);
-	}
-	else
-	{
-	    // Keep all other colors as is.
-	    translationtables[i] = translationtables[i+256] 
-		= translationtables[i+512] = i;
-	}
+        if (i >= 0x70 && i <= 0x7f)
+        {
+            // map green ramp to gray, brown, red
+            translationtables[i] = 0x60 + (i & 0xf);
+            translationtables[i + 256] = 0x40 + (i & 0xf);
+            translationtables[i + 512] = 0x20 + (i & 0xf);
+        }
+        else
+        {
+            // Keep all other colors as is.
+            translationtables[i] = translationtables[i + 256]
+                = translationtables[i + 512] = i;
+        }
     }
 }
